feat(config): add authentication configuration module
- Introduced `auth.py` to manage authentication settings.
- Updated `__init__.py` to include the new `auth` module.
- Refactored `legacy.py` to remove redundant authentication configurations.
