+++ backend/open_webui/config/__init__.py
@@ -3,6 +3,7 @@ from .logging import *
 from .database import *
 
 from .llm import *
+from .auth import *
 from .legacy import *
 
 setup_logging()
